website/pyproject.toml
[project]
name = "infolegAI-web"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite<0.22",
    "anyio>=4.12.0",
    "fastapi-users-db-sqlalchemy>=6.0.0",
    "fastapi-users[oauth]>=15.0.3",
    "fastapi[standard]>=0.125.0",
    "fastmcp>=2.14.3",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-community>=3.0.5",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langchain-pymupdf4llm>=0.5.0",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "sqlmodel>=0.0.27",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "ipython>=9.8.0",
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "rich>=14.2.0",
    "ruff>=0.14.10",
    "ty>=0.0.9",
]
